refactor types tests + add tests for process typedefs
