+++ test/types/process.ts
@@ -1,0 +1,17 @@
+/**
+ * Test extended type definitions of NodeJS Process interface
+ * @nuxt/vue-app/types/process.d.ts
+ */
+
+process.browser = true
+
+process.client = true
+
+process.mode = 'universal'
+process.mode = 'spa'
+
+process.modern = true
+
+process.server = true
+
+process.static = true
